backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Database
psycopg2-binary==2.9.11

# Authentication
djangorestframework-simplejwt==5.5.1
django-allauth==65.13.1
dj-rest-auth==7.0.1

# Environment variables
python-decouple==3.8

# Task scheduling and background jobs
celery==5.3.4
redis==5.0.1
django-celery-beat>=2.6.0

# Email
django-anymail==10.2

# File uploads
Pillow==10.1.0

# API documentation
drf-yasg==1.21.7


# Development
django-debug-toolbar==4.2.0

# ========== NEW DEPENDENCIES FOR FEATURES ==========

# WebSockets / Real-time (Feature 4)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# HTTP requests for integrations
requests==2.31.0

# Calendar integrations (Feature 8)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
msal==1.24.1

# Slack/Teams integrations (Feature 9)
slack-sdk==3.23.0

# GitHub integration (Feature 10)
PyGithub==2.1.1

# Voice command processing (Feature 6)
SpeechRecognition==3.10.0

# ML for predictions (Feature 7, 12)
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Sentiment analysis for burnout detection
textblob==0.17.1

